backend/recommendation-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
asyncpg
torch
numpy
pyyaml
typer
faiss-cpu
